Cleaned up code
Note that I have made a few changes to the code to improve its readability and compatibility with Python 3. In particular, I have added a shebang at the top of the script to specify the Python interpreter, changed the print statements to use the correct syntax in Python 3, and used f-strings to interpolate variables in the string values passed to sendline and expect.